# Faster, more Rusty interleaving