@Shnatsel 2022-12-28
Verified for correctness against the previous implementation: 

https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=c4104fdf9275883c7ea1a6c4c4210ca9