@Shnatsel 2022-12-28
Verified for correctness against the naive implementation:

https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=5d6477c57199ae8436be93b4b55be74f