@Shnatsel 2022-12-28
I've rewritten deinterleaving in a similar way, and validated it against the previous implementation as well.

Benchmarks:

Before:
```
test write_nonparallel_zip1_to_buffered      ... bench: 481,441,257 ns/iter (+/- 22,410,241)
test write_parallel_any_channels_to_buffered ... bench:  67,985,884 ns/iter (+/- 18,264,908)
test write_parallel_zip16_to_buffered        ... bench: 171,812,583 ns/iter (+/- 8,102,075)
test write_parallel_zip1_to_buffered         ... bench: 142,734,713 ns/iter (+/- 7,649,419)
test write_uncompressed_to_buffered          ... bench:  25,297,916 ns/iter (+/- 187,416)
```

After:
```
test write_nonparallel_zip1_to_buffered      ... bench: 437,675,887 ns/iter (+/- 8,042,909)
test write_parallel_any_channels_to_buffered ... bench:  50,899,408 ns/iter (+/- 5,720,138)
test write_parallel_zip16_to_buffered        ... bench: 163,239,642 ns/iter (+/- 19,723,059)
test write_parallel_zip1_to_buffered         ... bench: 130,192,218 ns/iter (+/- 7,385,720)
test write_uncompressed_to_buffered          ... bench:  23,991,998 ns/iter (+/- 90,758)
```

The gains are not as dramatic as on reads, but still noticeable.