@Shnatsel 2022-12-28
Upon closer inspection, `read_single_image_rle_all_channels` varies wildly, but is always faster on this branch.