@Shnatsel 2022-12-30
I'd appreciate if you could start pushing the "Merge" button on these PRs, so I could make further optimizations that build on them. So far I've been splitting them into different PRs that are all based on master, but now I find myself in situations where I need some PR already merged to make further changes (e.g. #176)